MVVM Pattern Added
