+++ FitnessApp/Home/View/SubView/ProgressCircleView.swift
@@ -1,0 +1,33 @@
+//
+//  ProgressCircleView.swift
+//  FitnessApp
+//
+//  Created by Maxim Dmitrochenko on 1/20/25.
+//
+
+import SwiftUI
+
+struct ProgressCircleView: View {
+    @Binding var progress: Int
+    var goal: Int
+    var color: Color
+    private let width: CGFloat = 20
+    
+    var body: some View {
+        ZStack {
+            Circle()
+                .stroke(color.opacity(0.3), lineWidth: width)
+            
+            Circle()
+                .rotation(.degrees(-90))
+                .trim(from: 0, to: CGFloat(progress) / CGFloat(goal))
+                .stroke(color, style: StrokeStyle(lineWidth: width, lineCap: .round))
+                .shadow(radius: 5)
+        }
+        .padding()
+    }
+}
+
+#Preview {
+    ProgressCircleView(progress: .constant(60), goal: 200, color: .red)
+}
